+++ tools/designer/designer/designerappiface.cpp
@@ -842,14 +842,14 @@ void DesignerFormWindowImpl::setImplementationIncludes( const QStringList &lst )
     for ( QStringList::ConstIterator sit = lst.begin(); sit != lst.end(); ++sit ) {
 	QString s = *sit;
 	if ( s.startsWith( "#include" ) )
-	    s.remove( 0, 8 );
+	    s.remove( (uint)0, 8 );
 	s = s.simplifyWhiteSpace();
 	if ( s[ 0 ] != '<' && s[ 0 ] != '"' ) {
 	    s.prepend( "\"" );
 	    s.append( "\"" );
 	}
 	if ( s[ 0 ] == '<' ) {
-	    s.remove( 0, 1 );
+	    s.remove( (uint)0, 1 );
 	    s.remove( s.length() - 1, 1 );
 	    MetaDataBase::Include inc;
 	    inc.header = s;
@@ -857,7 +857,7 @@ void DesignerFormWindowImpl::setImplementationIncludes( const QStringList &lst )
 	    inc.location = "global";
 	    includes << inc;
 	} else {
-	    s.remove( 0, 1 );
+	    s.remove( (uint)0, 1 );
 	    s.remove( s.length() - 1, 1 );
 	    MetaDataBase::Include inc;
 	    inc.header = s;
@@ -884,14 +884,14 @@ void DesignerFormWindowImpl::setDeclarationIncludes( const QStringList &lst )
     for ( QStringList::ConstIterator sit = lst.begin(); sit != lst.end(); ++sit ) {
 	QString s = *sit;
 	if ( s.startsWith( "#include" ) )
-	    s.remove( 0, 8 );
+	    s.remove( (uint)0, 8 );
 	s = s.simplifyWhiteSpace();
 	if ( s[ 0 ] != '<' && s[ 0 ] != '"' ) {
 	    s.prepend( "\"" );
 	    s.append( "\"" );
 	}
 	if ( s[ 0 ] == '<' ) {
-	    s.remove( 0, 1 );
+	    s.remove( (uint)0, 1 );
 	    s.remove( s.length() - 1, 1 );
 	    MetaDataBase::Include inc;
 	    inc.header = s;
@@ -899,7 +899,7 @@ void DesignerFormWindowImpl::setDeclarationIncludes( const QStringList &lst )
 	    inc.location = "global";
 	    includes << inc;
 	} else {
-	    s.remove( 0, 1 );
+	    s.remove( (uint)0, 1 );
 	    s.remove( s.length() - 1, 1 );
 	    MetaDataBase::Include inc;
 	    inc.header = s;
